ExerciseDetector/backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
